+++ dinosaurs.json
@@ -22,6 +22,9 @@
   {
     "type": "T-REX",
     "name": "Ranger Rex"
+  },
+  {
+    "type": "CHAD",
+    "name": "Kenny"
   }
-
 ]
