+++ prehistoric-creatures.json
@@ -16,11 +16,7 @@
     "img_url": "https://animalgeography.files.wordpress.com/2018/08/sloth-banner-e1535192925361.jpg?w=584&h=325"
   },
   {
-    "type": "Verified PC",
-    "img_url": "mac user"
-  },
-  {
-    "type": "CHAD",
-    "img_url": "Kenny.jpg"
+    "type": "Pacha",
+    "img_url": "https://i.kym-cdn.com/entries/icons/facebook/000/019/698/8RKAP94.jpg"
   }
 ]
