+++ views/dinosaurs/edit.ejs
@@ -1,9 +1,9 @@
   
 <form method="POST" action="/dinosaurs/<%= dinoId %>/?_method=PUT">
-    <label>Name</label>
+    <label>Name: </label>
     <input type="text" value="<%= dino.name %>" name="name" />
   
-    <label>Type</label>
+    <label>Type: </label>
     <input type="text" value="<%= dino.type %>" name="type" />
   
     <input type="submit" />
